вносит мелкие правки в компоненты
