+++ project/src/components/app/app.tsx
@@ -1,4 +1,4 @@
-import MainPage from '../pages/main-page/main-page';
+import MainPage from '../../pages/main-page';
 
 type PropsType = {
   placesCount: number,
