+++ project/src/components/card/card.tsx
@@ -34,5 +34,5 @@ export default function Card(): JSX.Element {
         <p className='place-card__type'>Apartment</p>
       </div>
     </article>
-  );
+  )
 }
